+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## 3.0.0 (25-06-2025)
+
+### Novità: 2
+- [#37577](https://parermine.regione.emilia-romagna.it/issues/37577) Kettle - Refactor migrazione a GCP mantenendo S3
+- [#37184](https://parermine.regione.emilia-romagna.it/issues/37184) Nuovi parametri per la connessione all'object storage
+
 ## 2.1.1 (10-12-2024)
 
 ### Bugfix: 1